requirements.txt
autogen-agentchat
autogen-ext[azure]
autogen_ext[openai]
python-dotenv
azure-identity
playwright